backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-docx
reportlab
requests
pydantic
python-multipart
aiofiles
jinja2
python-dotenv
sqlalchemy[asyncio]
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.2.*
stripe
razorpay
httpx
PyPDF2
